+++ umami/input_vars_tools/PlottingFunctions.py
@@ -4,18 +4,13 @@
 
 import os
 
-import matplotlib as mtp
-import matplotlib.pyplot as plt
 import numpy as np
-from matplotlib import gridspec
 
 import umami.data_tools as udt
 from umami.configuration import global_config, logger
-from umami.helper_tools import hist_ratio, hist_w_unc
 from umami.plotting import histogram, histogram_plot
 from umami.plotting.utils import translate_kwargs
 from umami.preprocessing_tools import GetVariableDict
-from umami.tools import applyATLASstyle, makeATLAStag
 
 
 def check_kwargs_var_plots(kwargs: dict, **custom_default):
@@ -26,38 +21,36 @@ def check_kwargs_var_plots(kwargs: dict, **custom_default):
     ----------
     kwargs: dict
         kwargs dictionary passed to the plotting functions
-        - plot_type : str
-            Plottype, like pdf or png
-        - UseAtlasTag : bool
+        - use_atlas_tag : bool
             Define if ATLAS Tag is used or not.
-        - ApplyATLASStyle : bool
+        - apply_atlas_style : bool
             Apply ATLAS Style of the plot (for approval etc.).
-        - AtlasTag : str
+        - atlas_first_tag : str
             Main tag. Mainly "Internal Simulation".
-        - SecondTag : str
+        - atlas_second_tag : str
             Lower tag in the ATLAS label with infos.
-        - yAxisAtlasTag : float
-            Y axis position of the ATLAS label.
-        - yAxisIncrease : float
+        - y_scale : float
             Y axis increase factor to fit the ATLAS label.
         - figsize : list
             List of the figure size. i.e [5, 6]
-        - Log : bool
+        - logy : bool
             Set y-axis log True or False.
         - ylabel : str
             Y-label.
-        - ycolor : str
-            Y-axis-label colour.
-        - legFontSize : int
+        - leg_fontsize : int
             Legend font size.
-        - ncol : int
+        - leg_ncol : int
             Number of columns of the legend.
-        - Bin_Width_y_axis : bool
+        - leg_loc : str
+            Location of the legend (matplotlib conventions)
+        - bin_width_in_ylabel : bool
             Option to specify the bin width in the ylabel, by default True
         - plot_type: str
             File format for the output, by default "pdf"
-        - transparent:
+        - transparent: bool
             Option to make the background of the plot transparent, by default True
+        - norm: bool
+            Normalise histograms, by default True
     **custom_default: dict
         overwrites the default values defined in this function
 
@@ -68,19 +61,18 @@ def check_kwargs_var_plots(kwargs: dict, **custom_default):
     """
     # the following kwargs all plotting functions have in common
     default_kwargs = {
-        "UseAtlasTag": True,
-        "ApplyATLASStyle": False,
-        "AtlasTag": "Internal Simulation",
-        "SecondTag": "$\\sqrt{s}$ = 13 TeV, $t\\bar{t}$ PFlow Jets",
-        "yAxisAtlasTag": 0.925,
-        "yAxisIncrease": 1,
-        "figsize": None,
-        "Log": True,
+        "use_atlas_tag": True,
+        "apply_atlas_style": True,
+        "atlas_first_tag": "Internal Simulation",
+        "atlas_second_tag": "$\\sqrt{s}$ = 13 TeV, $t\\bar{t}$ PFlow Jets",
+        "y_scale": 1,
+        "figsize": (6.8, 5),
+        "logy": True,
         "ylabel": "Number of Tracks",
-        "ycolor": "black",
-        "legFontSize": 10,
-        "ncol": 1,
-        "Bin_Width_y_axis": True,
+        "leg_fontsize": 10,
+        "leg_ncol": 1,
+        "leg_loc": "upper right",
+        "bin_width_in_ylabel": True,
         "plot_type": "pdf",
         "transparent": True,
         "norm": True,
@@ -95,7 +87,7 @@ def check_kwargs_var_plots(kwargs: dict, **custom_default):
         else:
             updated_kwargs[key] = value
 
-    # Add "normalised" to ylabel in case it was forgotten
+    # Add "Normalised" to ylabel in case it was forgotten
     if (
         updated_kwargs["norm"] is True
         and "norm" not in updated_kwargs["ylabel"].lower()
@@ -105,7 +97,7 @@ def check_kwargs_var_plots(kwargs: dict, **custom_default):
     return updated_kwargs
 
 
-def plot_nTracks_per_Jet(
+def plot_n_tracks_per_jet(
     datasets_filepaths: list,
     datasets_labels: list,
     datasets_track_names: list,
@@ -117,7 +109,9 @@ def plot_nTracks_per_Jet(
     **kwargs,
 ):
     """
-    Plot the number of tracks per jet as a histogram.
+    Plot the number of tracks per jet as a histogram. If multiple datasets are
+    provided, a ratio plot comparing each flavour individually will be added below
+    the main plot.
 
     Parameters
     ----------
@@ -141,6 +135,7 @@ def plot_nTracks_per_Jet(
         additional arguments passed to `check_kwargs_var_plots`
     """
     # check the kwargs
+    kwargs = translate_kwargs(kwargs)
     kwargs = check_kwargs_var_plots(kwargs)
     # Init Linestyles
     linestyles = ["solid", "dashed", "dotted", "dashdot"]
@@ -149,8 +144,7 @@ def plot_nTracks_per_Jet(
     trks_dict = {}
     flavour_label_dict = {}
 
-    # Iterate over the different dataset filepaths and labels
-    # defined in the config
+    # Iterate over the different dataset filepaths and labels defined in the config
     for (filepath, label, tracks_name) in zip(
         datasets_filepaths, datasets_labels, datasets_track_names
     ):
@@ -173,44 +167,40 @@ def plot_nTracks_per_Jet(
     logger.info(f"Path: {output_directory}")
     logger.info(f"Track origin: {track_origin}\n")
 
-    # Define the figure with two subplots of unequal sizes
-    axis_dict = {}
-
-    # Apply ATLAS style if true
-    if kwargs["ApplyATLASStyle"]:
-        applyATLASstyle(mtp)
-
-    # Set up new figure
-    if not kwargs["figsize"]:
-        fig = plt.figure(figsize=(11.69 * 0.8, 8.27 * 0.8))
-
-    else:
-        fig = plt.figure(figsize=(kwargs["figsize"][0], kwargs["figsize"][1]))
-
-    gs = gridspec.GridSpec(8, 1, figure=fig)
-    axis_dict["left"] = {}
-    axis_dict["left"]["top"] = fig.add_subplot(gs[:6, 0])
-    axis_dict["left"]["ratio"] = fig.add_subplot(
-        gs[6:, 0], sharex=axis_dict["left"]["top"]
+    # Initialise plot
+    n_tracks_plot = histogram_plot(
+        bins=np.arange(-0.5, 40.5, 1),
+        ymin_ratio_1=ratio_cut[0] if ratio_cut is not None else None,
+        ymax_ratio_1=ratio_cut[1] if ratio_cut is not None else None,
+        n_ratio_panels=1 if len(datasets_filepaths) > 1 else 0,
+        norm=kwargs["norm"],
+        y_scale=kwargs["y_scale"],
+        ylabel=kwargs["ylabel"],
+        logy=kwargs["logy"],
+        figsize=kwargs["figsize"],
+        leg_ncol=kwargs["leg_ncol"],
+        leg_loc=kwargs["leg_loc"],
+        apply_atlas_style=kwargs["apply_atlas_style"],
+        atlas_second_tag=kwargs["atlas_second_tag"],
+        bin_width_in_ylabel=kwargs["bin_width_in_ylabel"],
+    )
+    # Set xlabel
+    n_tracks_plot.xlabel = (
+        "Number of tracks per jet"
+        if track_origin == "All"
+        else f"Number of tracks per jet ({track_origin})"
     )
 
-    # Init bincounts for ratio calculation
-    bincounts = {}
-    bincounts_unc = {}
-
-    # Init lowest bincount
-    Lowest_histcount = 1
-
-    # Iterate over models
-    for model_number, (label, linestyle) in enumerate(
+    # Iterate over datasets
+    for dataset_number, (label, linestyle) in enumerate(
         zip(datasets_labels, linestyles[: len(datasets_labels)])
     ):
         # Sort after given variable
         trks = np.asarray(trks_dict[label])
         if track_origin == "All":
-            nTracks = np.sum(~np.isnan(trks["ptfrac"]), axis=1)
+            n_tracks = np.sum(~np.isnan(trks["ptfrac"]), axis=1)
         else:
-            nTracks = np.sum(
+            n_tracks = np.sum(
                 np.logical_and(
                     ~np.isnan(trks["ptfrac"]),
                     trks["truthOriginLabel"] == global_config.OriginType[track_origin],
@@ -218,194 +208,24 @@ def plot_nTracks_per_Jet(
                 axis=1,
             )
 
-        if model_number == 0:
-            # Calculate unified Binning
-            nTracks_first = nTracks[flavour_label_dict[label] == 0]
-
-            _, Binning = np.histogram(
-                nTracks_first,
-                bins=np.arange(-0.5, 40.5, 1),
-            )
-
         for flav_label, flavour in enumerate(class_labels):
-            nTracks_flavour = nTracks[flavour_label_dict[label] == flav_label]
 
-            # Calculate bins
-            bins, weights, unc, band = hist_w_unc(
-                a=nTracks_flavour,
-                bins=Binning,
+            n_tracks_flavour = n_tracks[flavour_label_dict[label] == flav_label]
+
+            n_tracks_plot.add(
+                histogram(
+                    values=n_tracks_flavour,
+                    flavour=flavour,
+                    label=label,
+                    linestyle=linestyle,
+                ),
+                reference=not bool(dataset_number),
             )
 
-            hist_counts, _, _ = axis_dict["left"]["top"].hist(
-                x=bins[:-1],
-                bins=bins,
-                weights=weights,
-                histtype="step",
-                linewidth=1.0,
-                linestyle=linestyle,
-                color=global_config.flavour_categories[flavour]["colour"],
-                stacked=False,
-                fill=False,
-                label=global_config.flavour_categories[flavour]["legend_label"]
-                + f" {label}",
-            )
-
-            if flavour == class_labels[-1] and model_number == 0:
-                axis_dict["left"]["top"].hist(
-                    x=bins[:-1],
-                    bins=bins,
-                    bottom=band,
-                    weights=unc * 2,
-                    label="stat. unc.",
-                    **global_config.hist_err_style,
-                )
-
-            else:
-                axis_dict["left"]["top"].hist(
-                    x=bins[:-1],
-                    bins=bins,
-                    bottom=band,
-                    weights=unc * 2,
-                    **global_config.hist_err_style,
-                )
-
-            bincounts.update({f"{flavour}{model_number}": hist_counts})
-            bincounts_unc.update({f"{flavour}{model_number}": unc})
-
-            for count in hist_counts:
-                if count != 0 and count < Lowest_histcount:
-                    Lowest_histcount = count
-
-        # Start ratio plot
-        if model_number != 0:
-            for flavour in class_labels:
-                step, step_unc = hist_ratio(
-                    numerator=bincounts[f"{flavour}{model_number}"],
-                    denominator=bincounts[f"{flavour}0"],
-                    numerator_unc=bincounts_unc[f"{flavour}{model_number}"],
-                    denominator_unc=bincounts_unc[f"{flavour}0"],
-                )
-
-                axis_dict["left"]["ratio"].step(
-                    x=Binning,
-                    y=step,
-                    color=global_config.flavour_categories[flavour]["colour"],
-                    linestyle=linestyles[model_number],
-                )
-
-                axis_dict["left"]["ratio"].fill_between(
-                    x=Binning,
-                    y1=step - step_unc,
-                    y2=step + step_unc,
-                    step="pre",
-                    facecolor="none",
-                    edgecolor=global_config.hist_err_style["edgecolor"],
-                    linewidth=global_config.hist_err_style["linewidth"],
-                    hatch=global_config.hist_err_style["hatch"],
-                )
-
-        elif model_number == 0:
-            # Add black line at one
-            axis_dict["left"]["ratio"].axhline(
-                y=1,
-                xmin=axis_dict["left"]["ratio"].get_xlim()[0],
-                xmax=axis_dict["left"]["ratio"].get_xlim()[1],
-                color="black",
-                alpha=0.5,
-            )
-
-    axis_dict["left"]["ratio"].set_xlim(
-        left=Binning[0],
-        right=Binning[-1],
-    )
-
-    # Add axes, titels and the legend
-    if kwargs["Bin_Width_y_axis"] is True:
-        Bin_Width = abs(Binning[1] - Binning[0])
-        axis_dict["left"]["top"].set_ylabel(
-            f"{kwargs['ylabel']} / {Bin_Width:.2f}",
-            fontsize=12,
-            horizontalalignment="right",
-            y=1.0,
-            color=kwargs["ycolor"],
-        )
-
-    else:
-        axis_dict["left"]["top"].set_ylabel(
-            kwargs["ylabel"],
-            fontsize=12,
-            horizontalalignment="right",
-            y=1.0,
-            color=kwargs["ycolor"],
-        )
-
-    axis_dict["left"]["top"].tick_params(axis="y", labelcolor=kwargs["ycolor"])
-
-    axis_dict["left"]["ratio"].set_xlabel(
-        "Number of tracks per Jet"
-        if track_origin == "All"
-        else f"Number of tracks per Jet ({track_origin})",
-        fontsize=12,
-        horizontalalignment="right",
-        x=1.0,
-    )
-
-    plt.setp(axis_dict["left"]["top"].get_xticklabels(), visible=False)
-
-    if kwargs["Log"] is True:
-        axis_dict["left"]["top"].set_yscale("log")
-
-        if axis_dict["left"]["top"].get_ylim()[0] <= 0:
-            # Set lower y limit
-            left_y_limits = axis_dict["left"]["top"].get_ylim()
-            axis_dict["left"]["top"].set_ylim(
-                bottom=Lowest_histcount * 0.1,
-                top=left_y_limits[1] * 10 ** (kwargs["yAxisIncrease"]),
-            )
-
-        else:
-            left_y_limits = axis_dict["left"]["top"].get_ylim()
-            axis_dict["left"]["top"].set_ylim(
-                bottom=left_y_limits[0] * 0.1,
-                top=left_y_limits[1] * 10 ** (kwargs["yAxisIncrease"]),
-            )
-
-    else:
-        left_y_limits = axis_dict["left"]["top"].get_ylim()
-        axis_dict["left"]["top"].set_ylim(
-            bottom=left_y_limits[0],
-            top=left_y_limits[1] * kwargs["yAxisIncrease"],
-        )
-
-    if ratio_cut is not None:
-        axis_dict["left"]["ratio"].set_ylim(bottom=ratio_cut[0], top=ratio_cut[1])
-
-    # Set axis
-    axis_dict["left"]["top"].legend(
-        loc="upper right",
-        fontsize=kwargs["legFontSize"],
-        ncol=kwargs["ncol"],
-    )
-
-    # Set tight layout
-    plt.tight_layout()
-
-    # Set ATLAS Tag
-    if kwargs["UseAtlasTag"] is True:
-        makeATLAStag(
-            ax=axis_dict["left"]["top"],
-            fig=fig,
-            first_tag=kwargs["AtlasTag"],
-            second_tag=kwargs["SecondTag"],
-            ymax=kwargs["yAxisAtlasTag"],
-        )
-
-    # Save and close figure
-    plt.savefig(
+    n_tracks_plot.draw()
+    n_tracks_plot.savefig(
         f"{output_directory}/nTracks_per_Jet_{track_origin}.{kwargs['plot_type']}"
     )
-    plt.close()
-    plt.clf()
 
 
 def plot_input_vars_trks(
@@ -424,7 +244,9 @@ def plot_input_vars_trks(
     **kwargs,
 ):
     """
-    Plot the track variable in comparison to another model with ratio plot.
+    Plot the track variable in comparison to another model with ratio plot. If multiple
+    datasets are provided, a ratio plot comparing each flavour individually will be
+    added below the main plot.
 
     Parameters
     ----------
@@ -463,8 +285,8 @@ def plot_input_vars_trks(
     # TODO: move remove ratio_cut and replace with "ratio_ymin" and "ratio_ymax" in
     # kwargs which are then handed to the histogram plot
     # check the kwargs
-    kwargs = check_kwargs_var_plots(kwargs)
     kwargs = translate_kwargs(kwargs)
+    kwargs = check_kwargs_var_plots(kwargs)
 
     # check to avoid dangerous default value (list)
     if n_leading is None:
@@ -518,8 +340,7 @@ def plot_input_vars_trks(
     trks_dict = {}
     flavour_label_dict = {}
 
-    # Iterate over the different dataset filepaths and labels
-    # defined in the config
+    # Iterate over the different dataset filepaths and labels defined in the config
     for filepath, label, tracks_name in zip(
         datasets_filepaths,
         datasets_labels,
@@ -611,19 +432,18 @@ def plot_input_vars_trks(
                 # Initialise plot for this variable
                 var_plot = histogram_plot(
                     bins=bins_dict[var],
-                    norm=kwargs["norm"],
                     ymin_ratio_1=ratio_cut[0] if ratio_cut is not None else None,
                     ymax_ratio_1=ratio_cut[1] if ratio_cut is not None else None,
+                    n_ratio_panels=1 if len(datasets_filepaths) > 1 else 0,
+                    norm=kwargs["norm"],
                     y_scale=kwargs["y_scale"],
                     ylabel=kwargs["ylabel"],
                     logy=kwargs["logy"],
-                    n_ratio_panels=1 if len(datasets_filepaths) > 1 else 0,
-                    figsize=(6.8, 5),
+                    figsize=kwargs["figsize"],
                     leg_ncol=kwargs["leg_ncol"],
-                    # TODO: use kwargs below when switching to True by default
-                    apply_atlas_style=True,
-                    # apply_atlas_style=kwargs["apply_atlas_style"],
+                    apply_atlas_style=kwargs["apply_atlas_style"],
                     atlas_second_tag=kwargs["atlas_second_tag"],
+                    bin_width_in_ylabel=kwargs["bin_width_in_ylabel"],
                 )
 
                 if n_lead is None:
@@ -638,8 +458,8 @@ def plot_input_vars_trks(
                         else (f"{n_lead+1} leading tracks {var} ({track_origin})")
                     )
 
-                # Iterate over models
-                for model_number, (label, linestyle) in enumerate(
+                # Iterate over datasets
+                for dataset_number, (label, linestyle) in enumerate(
                     zip(datasets_labels, linestyles[: len(datasets_labels)])
                 ):
                     # Sort after given variable
@@ -669,45 +489,34 @@ def plot_input_vars_trks(
 
                         # Get number of tracks
                         if track_origin == "All":
-                            Tracks = jets[:, n_lead][~np.isnan(jets[:, n_lead])]
+                            track_values = jets[:, n_lead][~np.isnan(jets[:, n_lead])]
                         else:
                             mask_nan = ~np.isnan(jets[var][:, n_lead])
                             mask_origin = np.asarray(
                                 jets[:, n_lead]["truthOriginLabel"]
                                 == global_config.OriginType[track_origin]
                             )
-                            Tracks = jets[:, n_lead][
+                            track_values = jets[:, n_lead][
                                 np.logical_and(mask_nan, mask_origin)
                             ][var]
 
                         # Add histogram to plot
                         var_plot.add(
                             histogram(
-                                values=Tracks,
+                                values=track_values,
                                 flavour=flavour,
                                 label=label,
                                 linestyle=linestyle,
                             ),
-                            reference=not bool(model_number),
+                            reference=not bool(dataset_number),
                         )
 
                 var_plot.draw()
-                if kwargs["bin_width_in_ylabel"] is True:
-                    bin_width = abs(var_plot.bins[1] - var_plot.bins[0])
-                    if bin_width < 1e-2:
-                        var_plot.ylabel = f"{kwargs['ylabel']} / {bin_width:.0e}"
-                    else:
-                        var_plot.ylabel = f"{kwargs['ylabel']} / {bin_width:.2f}"
-                    var_plot.set_ylabel(var_plot.axis_top)
-
                 var_plot.savefig(
                     f"{filedir}/{var}_{n_lead}_{track_origin}.{kwargs['plot_type']}",
                     transparent=kwargs["transparent"],
                 )
-        logger.info(
-            "\n-----------------------------------------------------------------------"
-            "--------"
-        )
+        logger.info(f"\n{80 * '-'}")
 
 
 def plot_input_vars_jets(
@@ -719,277 +528,13 @@ def plot_input_vars_jets(
     class_labels: list,
     special_param_jets: dict = None,
     output_directory: str = "input_vars_jets",
-    normalise: bool = True,
-    **kwargs,
-):
-    """
-    Plot the jet variable.
-
-    Parameters
-    ----------
-    datasets_filepaths: list
-        List of filepaths to the files.
-    datasets_labels: list
-        Label of the dataset for the legend.
-    var_dict: dict
-        Variable dict where all variables of the files are saved.
-    n_jets: int
-        Number of jets to use for plotting.
-    binning: dict
-        Decide which binning is used.
-    class_labels: list
-        List of class_labels which are to be plotted.
-    special_param_jets: dict
-        Give specific x-axis-limits for variable.
-    output_directory: str
-        Name of the output directory. Only the dir name not path!
-    normalise: bool, optional
-        Bool to specify if distributions are normalised, by default True.
-    **kwargs: dict
-        additional arguments passed to `check_kwargs_var_plots`
-
-    Raises
-    ------
-    ValueError
-        If the type of the given binning is not supported.
-    """
-
-    # Set the ylabel to jets
-    if "ylabel" not in kwargs:
-        kwargs["ylabel"] = "Number of Jets"
-
-    # check the kwargs
-    kwargs = check_kwargs_var_plots(kwargs, yAxisIncrease=10)
-
-    nBins_dict = {}
-
-    # Load the given binning or set it
-    for variable in binning:
-        if isinstance(binning[variable], list):
-            nBins_dict.update(
-                {
-                    variable: np.arange(
-                        binning[variable][0],
-                        binning[variable][1],
-                        binning[variable][2],
-                    )
-                }
-            )
-
-        # If int, set to the given numbers
-        elif isinstance(binning[variable], int):
-            nBins_dict.update({variable: binning[variable]})
-
-        # If None, give default value
-        elif binning[variable] is None:
-            nBins_dict.update({variable: int(100)})
-
-        else:
-            raise ValueError(f"Type {type(binning[variable])} is not supported!")
-
-    variable_config = GetVariableDict(var_dict)
-
-    # Init trks and flavour label dicts
-    jets_dict = {}
-    flavour_label_dict = {}
-
-    # Iterate over the different dataset filepaths and labels
-    # defined in the config
-    for (filepath, label) in zip(
-        datasets_filepaths,
-        datasets_labels,
-    ):
-
-        # Get the tracks and the labels from the file/files
-        jets, flavour_labels = udt.LoadJetsFromFile(
-            filepath=filepath,
-            class_labels=class_labels,
-            nJets=n_jets,
-            print_logger=False,
-        )
-
-        # Append jets to dict
-        jets_dict.update({label: jets})
-        flavour_label_dict.update({label: flavour_labels})
-
-    # Loading jet variables
-    jetsVars = [
-        i
-        for j in variable_config["train_variables"]
-        for i in variable_config["train_variables"][j]
-    ]
-
-    # Check if path is existing, if not mkdir
-    if not os.path.isdir(f"{output_directory}/"):
-        os.makedirs(f"{output_directory}/")
-    filedir = f"{output_directory}/"
-
-    # Loop over vars
-    for var in jetsVars:
-        if var in nBins_dict:
-            logger.info(f"Plotting {var}...")
-
-            for label in datasets_labels:
-                # Get variable and the labels of the jets
-                jets_var = jets_dict[label][var]
-                flavour_labels_var = flavour_label_dict[label]
-
-                # Clean both from nans
-                jets_var_clean = jets_var[~np.isnan(jets_var)]
-                flavour_label_clean = flavour_labels_var[~np.isnan(jets_var)]
-
-                # Calculate unified Binning
-                first_flav = jets_var_clean[flavour_label_clean == 0]
-
-                var_range = None
-                if special_param_jets is not None and var in special_param_jets:
-                    if (
-                        "lim_left" in special_param_jets[var]
-                        and "lim_right" in special_param_jets[var]
-                    ):
-                        lim_left = special_param_jets[var]["lim_left"]
-                        lim_right = special_param_jets[var]["lim_right"]
-                        var_range = (lim_left, lim_right)
-
-                if nBins_dict[var] is None:
-                    _, Binning = np.histogram(first_flav, range=var_range)
-
-                else:
-                    _, Binning = np.histogram(
-                        first_flav,
-                        bins=nBins_dict[var],
-                        range=var_range,
-                    )
-
-                # Apply ATLAS style if true
-                if kwargs["ApplyATLASStyle"]:
-                    applyATLASstyle(mtp)
-
-                # Set up new figure
-                if kwargs["figsize"] is None:
-                    fig = plt.figure(figsize=(11.69 * 0.8, 8.27 * 0.8))
-
-                else:
-                    fig = plt.figure(
-                        figsize=(kwargs["figsize"][0], kwargs["figsize"][1])
-                    )
-
-                for flav_label, flavour in enumerate(class_labels):
-                    jets_flavour = jets_var_clean[flavour_label_clean == flav_label]
-
-                    # Calculate bins
-                    bins, weights, unc, band = hist_w_unc(
-                        a=jets_flavour, bins=Binning, normed=normalise
-                    )
-
-                    plt.hist(
-                        x=bins[:-1],
-                        bins=bins,
-                        weights=weights,
-                        histtype="step",
-                        linewidth=1.0,
-                        color=global_config.flavour_categories[flavour]["colour"],
-                        stacked=False,
-                        fill=False,
-                        label=global_config.flavour_categories[flavour]["legend_label"],
-                    )
-
-                    if flavour == class_labels[-1]:
-                        plt.hist(
-                            x=bins[:-1],
-                            bins=bins,
-                            bottom=band,
-                            weights=unc * 2,
-                            label="stat. unc.",
-                            **global_config.hist_err_style,
-                        )
-
-                    else:
-                        plt.hist(
-                            x=bins[:-1],
-                            bins=bins,
-                            bottom=band,
-                            weights=unc * 2,
-                            **global_config.hist_err_style,
-                        )
-
-                plt.xlabel(var)
-
-                # Add axes, titels and the legend
-                if kwargs["Bin_Width_y_axis"] is True:
-                    Bin_Width = abs(Binning[1] - Binning[0])
-                    plt.ylabel(
-                        f"{kwargs['ylabel']} / {Bin_Width:.2f}",
-                        fontsize=12,
-                        horizontalalignment="right",
-                        y=1.0,
-                        color=kwargs["ycolor"],
-                    )
-
-                else:
-                    plt.ylabel(
-                        kwargs["ylabel"],
-                        fontsize=12,
-                        horizontalalignment="right",
-                        y=1.0,
-                        color=kwargs["ycolor"],
-                    )
-
-                if normalise is False:
-                    plt.ylabel(
-                        "Number of jets",
-                        fontsize=12,
-                        horizontalalignment="right",
-                        y=1.0,
-                        color=kwargs["ycolor"],
-                    )
-
-                if kwargs["Log"] is True:
-                    plt.yscale("log")
-                    ymin, ymax = plt.ylim()
-                    plt.ylim(ymin=0.01 * ymin, ymax=kwargs["yAxisIncrease"] * ymax)
-
-                else:
-                    ymin, ymax = plt.ylim()
-                    plt.ylim(ymin=0.8 * ymin, ymax=kwargs["yAxisIncrease"] * ymax)
-
-                plt.legend(
-                    loc="upper right",
-                    fontsize=kwargs["legFontSize"],
-                    ncol=kwargs["ncol"],
-                )
-                plt.tight_layout()
-
-                ax = plt.gca()
-                if kwargs["UseAtlasTag"] is True:
-                    makeATLAStag(
-                        ax=ax,
-                        fig=fig,
-                        first_tag=kwargs["AtlasTag"],
-                        second_tag=kwargs["SecondTag"],
-                        ymax=kwargs["yAxisAtlasTag"],
-                    )
-
-                plt.savefig(f"{filedir}/{var}.{kwargs['plot_type']}")
-                plt.close()
-                plt.clf()
-
-
-def plot_input_vars_jets_comparison(
-    datasets_filepaths: list,
-    datasets_labels: list,
-    var_dict: dict,
-    n_jets: int,
-    binning: dict,
-    class_labels: list,
-    special_param_jets: dict = None,
-    output_directory: str = "input_vars_jets",
     ratio_cut: list = None,
-    normalise: bool = True,
     **kwargs,
 ):
     """
-    Plot the jet variable comparison for the given datasets.
+    Plot the jet variable comparison for the given datasets. If multiple datasets are
+    provided, a ratio plot comparing each flavour individually will be added below
+    the main plot.
 
     Parameters
     ----------
@@ -1011,8 +556,6 @@ def plot_input_vars_jets_comparison(
         Name of the output directory. Only the dir name not path!
     ratio_cut : list
         List of y-axis cuts for the ratio block.
-    normalise: bool, optional
-        Bool to specify if distributions are normalised, by default True.
 
     **kwargs: dict
         additional arguments passed to `check_kwargs_var_plots`
@@ -1022,19 +565,21 @@ def plot_input_vars_jets_comparison(
     ValueError
         If the type of the given binning is not supported.
     """
+
+    # check the kwargs
+    kwargs = translate_kwargs(kwargs)
+    kwargs = check_kwargs_var_plots(kwargs, yAxisIncrease=10)
+
     # Set the ylabel to jets
     if "ylabel" not in kwargs:
         kwargs["ylabel"] = "Number of Jets"
 
-    # check the kwargs
-    kwargs = check_kwargs_var_plots(kwargs, yAxisIncrease=10)
-
-    nBins_dict = {}
+    bins_dict = {}
 
     # Load the given binning or set it
     for variable in binning:
         if isinstance(binning[variable], list):
-            nBins_dict.update(
+            bins_dict.update(
                 {
                     variable: np.arange(
                         binning[variable][0],
@@ -1046,11 +591,11 @@ def plot_input_vars_jets_comparison(
 
         # If int, set to the given numbers
         elif isinstance(binning[variable], int):
-            nBins_dict.update({variable: binning[variable]})
+            bins_dict.update({variable: binning[variable]})
 
         # If None, give default value
         elif binning[variable] is None:
-            nBins_dict.update({variable: int(100)})
+            bins_dict.update({variable: int(100)})
 
         else:
             raise ValueError(f"Type {type(binning[variable])} is not supported!")
@@ -1062,13 +607,8 @@ def plot_input_vars_jets_comparison(
     jets_dict = {}
     flavour_label_dict = {}
 
-    # Iterate over the different dataset filepaths and labels
-    # defined in the config
-    for (filepath, label) in zip(
-        datasets_filepaths,
-        datasets_labels,
-    ):
-
+    # Iterate over the different dataset filepaths and labels defined in the config
+    for (filepath, label) in zip(datasets_filepaths, datasets_labels):
         # Get the tracks and the labels from the file/files
         jets, flavour_labels = udt.LoadJetsFromFile(
             filepath=filepath,
@@ -1084,7 +624,7 @@ def plot_input_vars_jets_comparison(
     variable_config = GetVariableDict(var_dict)
 
     # Loading jet variables
-    jetsVars = [
+    jet_variables = [
         i
         for j in variable_config["train_variables"]
         for i in variable_config["train_variables"][j]
@@ -1096,40 +636,41 @@ def plot_input_vars_jets_comparison(
     filedir = f"{output_directory}/"
 
     # Loop over vars
-    for var in jetsVars:
-        if var in nBins_dict:
-            logger.info(f"Plotting {var}...")
+    for var in jet_variables:
+        if var in bins_dict:
 
-            # Define the figure with two subplots of unequal sizes
-            axis_dict = {}
-
-            # Apply ATLAS style if true
-            if kwargs["ApplyATLASStyle"]:
-                applyATLASstyle(mtp)
-
-            # Set up new figure
-            if kwargs["figsize"] is None:
-                fig = plt.figure(figsize=(11.69 * 0.8, 8.27 * 0.8))
-
-            else:
-                fig = plt.figure(figsize=(kwargs["figsize"][0], kwargs["figsize"][1]))
-
-            gs = gridspec.GridSpec(8, 1, figure=fig)
-            axis_dict["left"] = {}
-            axis_dict["left"]["top"] = fig.add_subplot(gs[:6, 0])
-            axis_dict["left"]["ratio"] = fig.add_subplot(
-                gs[6:, 0], sharex=axis_dict["left"]["top"]
+            # Initialise plot for this variable
+            var_plot = histogram_plot(
+                bins=bins_dict[var],
+                xlabel=var,
+                ymin_ratio_1=ratio_cut[0] if ratio_cut is not None else None,
+                ymax_ratio_1=ratio_cut[1] if ratio_cut is not None else None,
+                n_ratio_panels=1 if len(datasets_filepaths) > 1 else 0,
+                norm=kwargs["norm"],
+                y_scale=kwargs["y_scale"],
+                ylabel=kwargs["ylabel"],
+                logy=kwargs["logy"],
+                figsize=kwargs["figsize"],
+                leg_ncol=kwargs["leg_ncol"],
+                leg_loc=kwargs["leg_loc"],
+                apply_atlas_style=kwargs["apply_atlas_style"],
+                atlas_second_tag=kwargs["atlas_second_tag"],
+                bin_width_in_ylabel=kwargs["bin_width_in_ylabel"],
             )
+            # setting range based on value from config file
+            if special_param_jets is not None and var in special_param_jets:
+                if (
+                    "lim_left" in special_param_jets[var]
+                    and "lim_right" in special_param_jets[var]
+                ):
+                    lim_left = special_param_jets[var]["lim_left"]
+                    lim_right = special_param_jets[var]["lim_right"]
+                    var_plot.bins_range = (lim_left, lim_right)
 
-            # Init bincounts for ratio calculation
-            bincounts = {}
-            bincounts_unc = {}
+            logger.info(f"Plotting {var} ...")
 
-            # Init lowest bincount
-            Lowest_histcount = 1
-
-            # Iterate over models
-            for model_number, (label, linestyle) in enumerate(
+            # Iterate over datasets
+            for dataset_number, (label, linestyle) in enumerate(
                 zip(datasets_labels, linestyles[: len(datasets_labels)])
             ):
                 # Get variable and the labels of the jets
@@ -1140,215 +681,25 @@ def plot_input_vars_jets_comparison(
                 jets_var_clean = jets_var[~np.isnan(jets_var)]
                 flavour_label_clean = flavour_labels_var[~np.isnan(jets_var)]
 
-                if model_number == 0:
-                    # Calculate unified Binning
-                    first_flav = jets_var_clean[flavour_label_clean == 0]
-
-                    var_range = None
-                    if special_param_jets is not None and var in special_param_jets:
-                        if (
-                            "lim_left" in special_param_jets[var]
-                            and "lim_right" in special_param_jets[var]
-                        ):
-                            lim_left = special_param_jets[var]["lim_left"]
-                            lim_right = special_param_jets[var]["lim_right"]
-                            var_range = (lim_left, lim_right)
-
-                    if nBins_dict[var] is None:
-                        _, Binning = np.histogram(first_flav, range=var_range)
-
-                    else:
-                        _, Binning = np.histogram(
-                            first_flav,
-                            bins=nBins_dict[var],
-                            range=var_range,
-                        )
-
                 for flav_label, flavour in enumerate(class_labels):
                     jets_flavour = jets_var_clean[flavour_label_clean == flav_label]
 
-                    # Calculate bins
-                    bins, weights, unc, band = hist_w_unc(
-                        a=jets_flavour, bins=Binning, normed=normalise
+                    # Add histogram to plot
+                    var_plot.add(
+                        histogram(
+                            values=jets_flavour,
+                            flavour=flavour,
+                            label=label,
+                            linestyle=linestyle,
+                        ),
+                        reference=not bool(dataset_number),
                     )
 
-                    hist_counts, _, _ = axis_dict["left"]["top"].hist(
-                        x=bins[:-1],
-                        bins=bins,
-                        weights=weights,
-                        histtype="step",
-                        linewidth=1.0,
-                        linestyle=linestyle,
-                        color=global_config.flavour_categories[flavour]["colour"],
-                        stacked=False,
-                        fill=False,
-                        label=global_config.flavour_categories[flavour]["legend_label"]
-                        + f" {label}",
-                    )
-
-                    if flavour == class_labels[-1] and model_number == 0:
-                        axis_dict["left"]["top"].hist(
-                            x=bins[:-1],
-                            bins=bins,
-                            bottom=band,
-                            weights=unc * 2,
-                            label="stat. unc.",
-                            **global_config.hist_err_style,
-                        )
-
-                    else:
-                        axis_dict["left"]["top"].hist(
-                            x=bins[:-1],
-                            bins=bins,
-                            bottom=band,
-                            weights=unc * 2,
-                            **global_config.hist_err_style,
-                        )
-
-                    bincounts.update({f"{flavour}{model_number}": hist_counts})
-                    bincounts_unc.update({f"{flavour}{model_number}": unc})
-
-                    for count in hist_counts:
-                        if count != 0 and count < Lowest_histcount:
-                            Lowest_histcount = count
-
-                # Start ratio plot
-                if model_number != 0:
-                    for flavour in class_labels:
-                        step, step_unc = hist_ratio(
-                            numerator=bincounts[f"{flavour}{model_number}"],
-                            denominator=bincounts[f"{flavour}0"],
-                            numerator_unc=bincounts_unc[f"{flavour}{model_number}"],
-                            denominator_unc=bincounts_unc[f"{flavour}0"],
-                        )
-
-                        axis_dict["left"]["ratio"].step(
-                            x=Binning,
-                            y=step,
-                            color=global_config.flavour_categories[flavour]["colour"],
-                            linestyle=linestyles[model_number],
-                        )
-
-                        axis_dict["left"]["ratio"].fill_between(
-                            x=Binning,
-                            y1=step - step_unc,
-                            y2=step + step_unc,
-                            step="pre",
-                            facecolor="none",
-                            edgecolor=global_config.hist_err_style["edgecolor"],
-                            linewidth=global_config.hist_err_style["linewidth"],
-                            hatch=global_config.hist_err_style["hatch"],
-                        )
-
-                elif model_number == 0:
-                    # Add black line at one
-                    axis_dict["left"]["ratio"].axhline(
-                        y=1,
-                        xmin=axis_dict["left"]["ratio"].get_xlim()[0],
-                        xmax=axis_dict["left"]["ratio"].get_xlim()[1],
-                        color="black",
-                        alpha=0.5,
-                    )
-
-            axis_dict["left"]["ratio"].set_xlim(
-                left=Binning[0],
-                right=Binning[-1],
+            # Draw and save the plot
+            var_plot.draw()
+            var_plot.savefig(
+                f"{filedir}/{var}.{kwargs['plot_type']}",
+                transparent=kwargs["transparent"],
             )
 
-            axis_dict["left"]["ratio"].set_ylabel("Ratio", fontsize=12)
-
-            # Add axes, titels and the legend
-            if kwargs["Bin_Width_y_axis"] is True:
-                Bin_Width = abs(Binning[1] - Binning[0])
-                axis_dict["left"]["top"].set_ylabel(
-                    f"{kwargs['ylabel']} / {Bin_Width:.2f}",
-                    fontsize=12,
-                    horizontalalignment="right",
-                    y=1.0,
-                    color=kwargs["ycolor"],
-                )
-
-            else:
-                axis_dict["left"]["top"].set_ylabel(
-                    kwargs["ylabel"],
-                    fontsize=12,
-                    horizontalalignment="right",
-                    y=1.0,
-                    color=kwargs["ycolor"],
-                )
-
-            if normalise is False:
-                axis_dict["left"]["top"].set_ylabel(
-                    "Number of jets",
-                    fontsize=12,
-                    horizontalalignment="right",
-                    y=1.0,
-                    color=kwargs["ycolor"],
-                )
-
-            axis_dict["left"]["top"].tick_params(axis="y", labelcolor=kwargs["ycolor"])
-
-            axis_dict["left"]["ratio"].set_xlabel(
-                var, fontsize=12, horizontalalignment="right", x=1.0
-            )
-
-            plt.setp(axis_dict["left"]["top"].get_xticklabels(), visible=False)
-
-            if kwargs["Log"] is True:
-                axis_dict["left"]["top"].set_yscale("log")
-
-                if axis_dict["left"]["top"].get_ylim()[0] <= 0:
-                    # Set lower y limit
-                    left_y_limits = axis_dict["left"]["top"].get_ylim()
-                    axis_dict["left"]["top"].set_ylim(
-                        bottom=Lowest_histcount * 0.1,
-                        top=left_y_limits[1] * 10 ** (kwargs["yAxisIncrease"]),
-                    )
-
-                else:
-                    left_y_limits = axis_dict["left"]["top"].get_ylim()
-                    axis_dict["left"]["top"].set_ylim(
-                        bottom=left_y_limits[0] * 0.1,
-                        top=left_y_limits[1] * 10 ** (kwargs["yAxisIncrease"]),
-                    )
-
-            else:
-                left_y_limits = axis_dict["left"]["top"].get_ylim()
-                axis_dict["left"]["top"].set_ylim(
-                    bottom=left_y_limits[0],
-                    top=left_y_limits[1] * kwargs["yAxisIncrease"],
-                )
-
-            if ratio_cut is not None:
-                axis_dict["left"]["ratio"].set_ylim(
-                    bottom=ratio_cut[0], top=ratio_cut[1]
-                )
-
-            # Set axis
-            axis_dict["left"]["top"].legend(
-                loc="upper right",
-                fontsize=kwargs["legFontSize"],
-                ncol=kwargs["ncol"],
-            )
-
-            # Set tight layout
-            plt.tight_layout()
-
-            # Set ATLAS Tag
-            if kwargs["UseAtlasTag"] is True:
-                makeATLAStag(
-                    ax=axis_dict["left"]["top"],
-                    fig=fig,
-                    first_tag=kwargs["AtlasTag"],
-                    second_tag=kwargs["SecondTag"],
-                    ymax=kwargs["yAxisAtlasTag"],
-                )
-
-            # Save and close figure
-            plt.savefig(f"{filedir}/{var}.{kwargs['plot_type']}")
-            plt.close()
-            plt.clf()
-    logger.info(
-        "\n------------------------------------------------------------------"
-        "-------------"
-    )
+    logger.info(f"\n{80 * '-'}")