+++ umami/input_vars_tools/__init__.py
@@ -3,7 +3,6 @@
 from umami.input_vars_tools.PlottingFunctions import (
     check_kwargs_var_plots,
     plot_input_vars_jets,
-    plot_input_vars_jets_comparison,
     plot_input_vars_trks,
-    plot_nTracks_per_Jet,
+    plot_n_tracks_per_jet,
 )